rename eventlog to eventslog
