+++ frontend/src/modules/events-log/EventType.tsx
