+++ frontend/src/modules/events-log/constants.ts
