+++ frontend/src/modules/events-log/container/EventsLogDetails.tsx
